LMA-353 - > 311 Web_User profile problem SubTask # 4: Set default Industry field value to "Web Developement"
Code Refactoring: Set industry field as not nullable and empty, then set the default industry value. In order to prevent nullable industry field, we created a fixture file for course_overview categories, this fixture would be required to load (using loaddata) when database server is migrated or being newly setup on anyother machine.