fix: explicitly allow courseware media access in HTTP response
Set the experimental HTTP Feature-Policy response header to
allow content rendered by (i) the chromeless xblock view and
(ii) the legacy courseware index to request access to mic,
camera, MIDI, location, and encrpyted media.

Access to these features was implicitly allowed in older
browser versions. However, in the newest version of
Chromium and in near-future versions of Firefox,
ability for content to request media access will require
that media to be explicitly allowed in the HTTP response
headers.

TNL-7675